Update deploy controls
