+++ lit-api-server/blockchain/lit_node_express/Makefile
@@ -7,7 +7,6 @@ DEPLOYER_BIN := $(RUST_PROJECT)/target/debug/contract_deployer
 DEPLOYER_SECRET := 0xd9884c772d9e85843d36fc4b18972068efc02ec6f50924725d2cc39ecefcd8ef
 BASE_DEPLOYER_SECRET := 0x747ea873f95c08f90634f496bbdbf53b852e297704499dfd3494e3046d44f016
 ANVIL_SECRET := 0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
-DIAMOND_ADDRESS := 0x17788103ca8f9bc39c43b70b601dac99be5a63cb
 
 # 1. Compile Solidity contracts with Hardhat
 compile:
@@ -40,8 +39,8 @@ deploy_anvil: compile
 
 update_anvil: compile
 	cd $(RUST_PROJECT) && cargo build --bin contract_deployer
-	$(DEPLOYER_BIN) --action=update --network=anvil --abifolder=artifacts/contracts --secret=$(ANVIL_SECRET) --address=$(DIAMOND_ADDRESS)
+	$(DEPLOYER_BIN) --action=update --network=anvil --abifolder=artifacts/contracts --secret=$(ANVIL_SECRET) --address=$(address)
 
 update_sepolia: compile
 	cd $(RUST_PROJECT) && cargo build --bin contract_deployer
-	$(DEPLOYER_BIN) --action=update --network=base-sepolia --abifolder=artifacts/contracts --secret=$(DEPLOYER_SECRET) --address=$(DIAMOND_ADDRESS)
+	$(DEPLOYER_BIN) --action=update --network=base-sepolia --abifolder=artifacts/contracts --secret=$(DEPLOYER_SECRET) --address=$(address)
