+++ lit-api-server/blockchain/rust_generator_and_deployer/src/deployer/mod.rs
@@ -1,0 +1,2 @@
+pub mod deploy;
+pub mod diamond;
