+++ lit-api-server/blockchain/rust_generator_and_deployer/src/lib.rs
@@ -1,2 +1,3 @@
-pub (crate) mod diamond;
-pub (crate) mod common;
+pub mod args;
+pub mod deployer;
+pub mod diamond;
